Move schema-parser.js to lib
