+++ src/lib/schema-parser.js
@@ -1,9 +1,9 @@
 import $RefParser from 'json-schema-ref-parser';
-import { getAttributes, getRelationships } from './lib/normalize';
-import { request } from './lib/request';
-import { extract } from './utils';
-import Document from './lib/document';
-import { compileJsonApiUrl, parseJsonApiUrl } from './lib/url';
+import { getAttributes, getRelationships } from './normalize';
+import { request } from './request';
+import { extract } from '../utils';
+import Document from './document';
+import { compileJsonApiUrl, parseJsonApiUrl } from './url';
 
 export default class SchemaParser {
   constructor() {
